[Database] Uhhh several things again
