+++ Zebra/Model/ZBSource.m
@@ -13,10 +13,7 @@
 #import "ZBSourceManager.h"
 #import "UICKeyChainStore.h"
 #import <ZBAppDelegate.h>
-//#import <Database/ZBDatabaseManager.h>
-#import <Database/ZBColumn.h>
 #import <ZBDevice.h>
-#import <ZBUtils.h>
 #import <JSONParsing/ZBUserInfo.h>
 #import <JSONParsing/ZBSourceInfo.h>
 #import <Tabs/ZBTabBarController.h>
