+++ Zebra/Resources/default.list
@@ -1,1 +1,2 @@
 # Zebra Sources List
+deb https://getzbra.com/repo/ ./
