+++ Zebra/Tabs/Home/Wish List/ZBWishListTableViewController.m
@@ -13,7 +13,6 @@
 #import <ZBSettings.h>
 #import <Queue/ZBQueue.h>
 #import <ZBDevice.h>
-#import <Managers/Delegates/ZBDatabaseDelegate.h>
 #import <Extensions/UINavigationBar+Extensions.h>
 #import <Extensions/UIColor+GlobalColors.h>
 #import <UI/Views/Cells/ZBPackageTableViewCell.h>
