+++ Zebra/Tabs/Packages/Controllers/ZBPackagesByAuthorTableViewController.m
@@ -6,6 +6,8 @@
 //  Copyright © 2019 Wilson Styres. All rights reserved.
 //
 
+// I'm going to replace this class with a ZBPackageListController later so I'm not going to worry about it too much for now
+
 #import "ZBPackagesByAuthorTableViewController.h"
 
 #import <ZBAppDelegate.h>
@@ -18,8 +20,8 @@
 
 @interface ZBPackagesByAuthorTableViewController () {
     NSArray *moreByAuthor;
+    ZBPackage *package;
 }
-@property (nonatomic, strong) ZBPackage *package;
 @end
 
 @implementation ZBPackagesByAuthorTableViewController
@@ -28,7 +30,7 @@
     self = [super init];
     
     if (self) {
-        self.package = package;
+        self->package = package;
     }
     
     return self;
@@ -36,9 +38,9 @@
 
 - (void)viewDidLoad {
     [super viewDidLoad];
-//    moreByAuthor = [[ZBDatabaseManager sharedInstance] packagesByAuthorName:self.package.authorName email:self.package.authorEmail];
+    moreByAuthor = [[ZBDatabaseManager sharedInstance] packagesByAuthorWithName:package.authorName email:package.authorEmail];
     [self.tableView registerNib:[UINib nibWithNibName:@"ZBPackageTableViewCell" bundle:nil] forCellReuseIdentifier:@"packageTableViewCell"];
-    self.navigationItem.title = self.package.authorName;
+    self.navigationItem.title = package.authorName;
     self.tableView.tableFooterView = [[UIView alloc] initWithFrame:CGRectZero];
 }
 
@@ -66,12 +68,12 @@
 }
 
 - (void)tableView:(UITableView *)tableView willDisplayCell:(UITableViewCell *)cell forRowAtIndexPath:(NSIndexPath *)indexPath {
-    ZBPackage *package = [[ZBDatabaseManager sharedInstance] topVersionForPackage:moreByAuthor[indexPath.row]];
+    ZBPackage *package = moreByAuthor[indexPath.row];
     [(ZBPackageTableViewCell *)cell updateData:package];
 }
 
 - (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath {
-    ZBPackage *package = [[ZBDatabaseManager sharedInstance] topVersionForPackage:[moreByAuthor objectAtIndex:indexPath.row]];
+    ZBPackage *package = moreByAuthor[indexPath.row];
     if (package) {
         ZBPackageViewController *packageDepiction = [[ZBPackageViewController alloc] initWithPackage:package];
